fix: expose Tool.failureResultSchema and widen tool failure result types
Move HttpRequestDetails and HttpResponseDetails to AiError so Response can
import Tool without a cycle. Order the failure result union as AiError,
the tool failure schema, then ExecutionFailure so user failures keep their
fields. Remove accidentally committed generated docs.
