--- packages/effect/docs/_config.yml
@@ -1,9 +1,0 @@
-remote_theme: mikearnaldi/just-the-docs
-
-# Enable or disable the site search
-search_enabled: true
-
-# Aux links for the upper right navigation
-aux_links:
-'Homepage':
-  - 'https://effect.website'
