--- packages/effect/docs/index.md
@@ -1,4 +1,0 @@
----
-title: Home
-nav_order: 1
----
