--- packages/effect/docs/modules/index.md
@@ -1,6 +1,0 @@
----
-title: Modules
-has_children: true
-permalink: /docs/modules
-nav_order: 2
----
